parser (core) is now closure-bound
No need to do cleanup in every test as "cli" now requires instantiation
and tracks all its state per instance. The API does not change; when the
user requires "cliargs" they will be provided with a default instance of
`cliargs.core`.

TODO: restore the functionality of automatic cleanup after parse.
