+++ spec/utils/wordwrap_spec.lua
@@ -1,4 +1,4 @@
-local subject = require('utils/wordwrap')
+local subject = require('cliargs.utils.wordwrap')
 
 describe("utils::wordwrap", function()
   it("should work", function()
